refactor(runtime): 收口 remote 与 wav 运行时入口
Why:
- 将 remote 与 wav 统一归入 runtime 运行期支撑层，降低顶层目录语义混杂
- 收掉 unpack 侧旧 WAV bridge，统一主流程对运行时能力的入口

What:
- 新增 runtime 顶层导出面，并将 remote 无损迁移到 runtime/remote
- 将 wav 真实实现收为 runtime/wav 子包，保留根级 remote/wav 兼容 shim
- 更新 facade、cli dispatch、unpack batch 与相关测试到新的 runtime 入口，并删除 unpack/wav.py

Risk:
- 兼容性影响：有，移除了 lol_audio_unpack.unpack.wav 旧路径；根级 remote/wav 入口仍保留兼容层
- 潜在风险：若仓库外部仍直接依赖旧 unpack.wav 或后台 WAV 模块路径，需要同步迁移到 runtime 入口

Test:
- uv run pytest -q tests/test_remote_package_api.py tests/test_remote_preparer.py tests/test_context_injection.py tests/test_runtime_wav_package_api.py tests/test_wav_sidecar.py tests/test_unpack_wav_sidecar.py tests/test_cli.py tests/test_task_runner.py -> 84 passed in 18.44s
- uv run ruff check src/lol_audio_unpack/runtime src/lol_audio_unpack/facade.py src/lol_audio_unpack/cli/dispatch.py src/lol_audio_unpack/gui/service/task_runner.py src/lol_audio_unpack/unpack/batch.py tests/test_remote_package_api.py tests/test_remote_preparer.py tests/test_runtime_wav_package_api.py tests/test_wav_sidecar.py tests/test_unpack_wav_sidecar.py tests/test_cli.py tests/test_task_runner.py -> All checks passed!
- uv run python -m py_compile src/lol_audio_unpack/runtime/__init__.py src/lol_audio_unpack/runtime/remote/__init__.py src/lol_audio_unpack/runtime/wav/__init__.py src/lol_audio_unpack/runtime/wav/background_job.py src/lol_audio_unpack/runtime/wav/coordinator.py src/lol_audio_unpack/runtime/wav/_runtime.py src/lol_audio_unpack/facade.py src/lol_audio_unpack/cli/dispatch.py src/lol_audio_unpack/gui/service/task_runner.py src/lol_audio_unpack/unpack/batch.py -> pass
