+++ src/lol_audio_unpack/runtime/wav/__init__.py
@@ -1,27 +1,38 @@
-"""兼容层：迁移期保留旧后台 WAV 作业导入路径。"""
+"""WAV 运行期能力公开导出面。"""
 
-from lol_audio_unpack.runtime.wav.background_job import (
+from ._runtime import (
+    AttemptResult,
+    WavJob,
+    WavJobFailure,
+    WavSidecarSummary,
+    build_wav_output_path,
+    default_worker_entry,
+    resolve_wav_decode_config,
+)
+from .background_job import (
     WavBackgroundJobSpec,
     WavBackgroundProcessHandle,
     WavManifestRecorder,
     build_manifest_recorder,
-    build_wav_background_job_spec_from_args,
     build_wav_background_job_spec_from_paths,
     launch_detached_wav,
     launch_wav_background_process,
-    main,
     run_wav_background_job,
 )
+from .coordinator import WavSidecarProgressSnapshot, WavTranscodeCoordinator
 
 __all__ = [
     "WavBackgroundJobSpec",
     "WavBackgroundProcessHandle",
     "WavManifestRecorder",
+    "WavSidecarProgressSnapshot",
+    "WavSidecarSummary",
+    "WavTranscodeCoordinator",
     "build_manifest_recorder",
-    "build_wav_background_job_spec_from_args",
     "build_wav_background_job_spec_from_paths",
+    "build_wav_output_path",
     "launch_detached_wav",
     "launch_wav_background_process",
-    "main",
+    "resolve_wav_decode_config",
     "run_wav_background_job",
 ]
